refactor: modularize server architecture for improved maintainability
- Break down monolithic 923-line index.ts into focused modules
- Extract tool definitions into dedicated tool-definitions.ts module
- Move all tool handlers to separate tool-handlers.ts module
- Create object-creators.ts for unified object creation logic
- Implement ServerManager class for clean server lifecycle management
- Reduce index.ts from 923 to 77 lines (~92% reduction)
- Improve code organization following Single Responsibility Principle
- Maintain full backward compatibility with existing functionality
- Enable better unit testing and module reusability

Files changed:
- src/index.ts: Simplified to startup orchestration only
- src/modules/server-manager.ts: New server lifecycle management
- src/modules/tool-definitions.ts: New centralized tool schema definitions
- src/modules/tool-handlers.ts: New consolidated tool request handlers
- src/modules/object-creators.ts: New unified object creation system

This refactoring improves maintainability while preserving all existing
functionality and MCP protocol compliance.
